docs: fix highlighting (code:: -> code-block::)
Some parts of the docs were using a ReST code:: directive to demarcate
code samples.  I can't find this directive in the docs; it seems to
mostly work, but it doesn't trigger syntax highlighting.

In every case, it seems like the author *wanted* syntax highlighting, so
I've replaced these directives with the more standard code-block::.